requirements.txt
Django==3.2.8
django-allauth==0.46.0
django-cors-headers==3.10.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
requests==2.26.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
python-dateutil==2.8.2
